LayoutTests/fast/custom-elements/upgrading/upgrading-parser-created-element.html

<!DOCTYPE html>
<html>
<head>
<title>Custom Elements: Upgrading unresolved elements</title>
<meta name="author" title="Ryosuke Niwa" href="mailto:rniwa@webkit.org">
<meta name="assert" content="HTML parser must add an unresolved custom element to the upgrade candidates map">
<link rel="help" href="https://w3c.github.io/webcomponents/spec/custom/#upgrading">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link rel='stylesheet' href='../../../resources/testharness.css'>
</head>
<body>
<div id="log"></div>
<my-custom-element></my-custom-element>
<instantiates-itself-after-super></instantiates-itself-after-super>
<instantiates-itself-before-super></instantiates-itself-before-super>
<my-other-element id="instance"></my-other-element>
<my-other-element id="otherInstance"></my-other-element>
<script>

test(function () {
    class MyCustomElement extends HTMLElement { }

    var instance = document.querySelector('my-custom-element');
    assert_true(instance instanceof HTMLElement);
    assert_false(instance instanceof HTMLUnknownElement,
        'an unresolved custom element should not be an instance of HTMLUnknownElement');
    assert_false(instance instanceof MyCustomElement);

    document.defineElement('my-custom-element', MyCustomElement);

    assert_true(instance instanceof HTMLElement);
    assert_true(instance instanceof MyCustomElement,
        'Calling defineElement must upgrade existing custom elements');

}, 'Element.prototype.createElement must add an unresolved custom element to the upgrade candidates map');


test(function () {
    class InstantiatesItselfAfterSuper extends HTMLElement {
        constructor(doNotCreateItself) {
            super();
            if (!doNotCreateItself)
                new InstantiatesItselfAfterSuper(true);
        }
    }

    assert_throws({'name': 'InvalidStateError'}, function () {
        document.defineElement('instantiates-itself-after-super', InstantiatesItselfAfterSuper);
    });
}, 'HTMLElement constructor must throw an InvalidStateError when the top of the construction stack is marked AlreadyConstructed'
    + ' due to a custom element constructor constructing itself after super() call');

test(function () {
    class InstantiatesItselfBeforeSuper extends HTMLElement {
        constructor(doNotCreateItself) {
            if (!doNotCreateItself)
                new InstantiatesItselfBeforeSuper(true);
            super();
        }
    }

    assert_throws({'name': 'InvalidStateError'}, function () {
        document.defineElement('instantiates-itself-before-super', InstantiatesItselfBeforeSuper);
    });
}, 'HTMLElement constructor must throw an InvalidStateError when the top of the construction stack is marked AlreadyConstructed'
    + ' due to a custom element constructor constructing itself before super() call');

test(function () {
    class MyOtherElement extends HTMLElement {
        constructor() {
            super();
            if (this == instance)
                return otherInstance;
        }
    }
    var instance = document.getElementById('instance');
    var otherInstance = document.getElementById('otherInstance');

    assert_false(instance instanceof MyOtherElement);
    assert_false(otherInstance instanceof MyOtherElement);

    assert_throws({'name': 'InvalidStateError'}, function () {
        document.defineElement('my-other-element', MyOtherElement);
    });

    assert_true(document.createElement('my-other-element') instanceof MyOtherElement,
        'Upgrading of custom elements must happen after the definition was added to the registry.');
}, 'Upgrading a custom element must throw an InvalidStateError when the returned element is not SameValue as the upgraded element');

</script>
</body>
</html>
